Complete Code

+++ MCRoomBookUI/src/app/admin/rooms/room-detail/room-detail.component.ts
@@ -1,4 +1,4 @@
-import {Component, Input, OnInit} from '@angular/core';
+import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 import {Room} from '../../../model/Room';
 import {Router} from '@angular/router';
 import {DataService} from '../../../data.service';
@@ -13,6 +13,11 @@ export class RoomDetailComponent implements OnInit {
   @Input()
   room: Room;
 
+  @Output()
+  dataChangedEvent = new EventEmitter();
+
+  message = '';
+
   constructor(private router: Router,
               private dataService: DataService) { }
 
@@ -24,9 +29,18 @@ export class RoomDetailComponent implements OnInit {
   }
 
   deleteRoom() {
-    this.dataService.deleteRoom(this.room.id).subscribe(
-      next => this.router.navigate(['admin','rooms'])
-    );
+    const result = confirm('Are you sure you wish to delete this room?');
+    if (result) {
+      this.message = 'deleting...';
+      this.dataService.deleteRoom(this.room.id).subscribe(
+        next => {
+          this.dataChangedEvent.emit();
+          this.router.navigate(['admin', 'rooms']);
+        }, error => {
+          this.message = 'Sorry - this room cannot be deleted at this time.';
+        }
+      );
+    }
   }
 
 }
