+++ MCRoomBookUI/src/app/admin/users/user-detail/user-detail.component.ts
@@ -1,4 +1,4 @@
-import {Component, Input, OnInit} from '@angular/core';
+import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 import {User} from '../../../model/User';
 import {DataService} from '../../../data.service';
 import {Router} from '@angular/router';
@@ -13,6 +13,11 @@ export class UserDetailComponent implements OnInit {
   @Input()
   user: User;
 
+  @Output()
+  dataChangedEvent = new EventEmitter();
+
+  message = '';
+
   constructor(private dataService: DataService,
               private router: Router) { }
 
@@ -25,15 +30,24 @@ export class UserDetailComponent implements OnInit {
   }
 
   deleteUser() {
-    this.dataService.deleteUser(this.user.id).subscribe(
-      next => {
-        this.router.navigate(['admin','users']);
-      }
-    );
+    const result = confirm('Are you sure you wish to delete this user?');
+    if (result) {
+      this.message = 'deleting...';
+      this.dataService.deleteUser(this.user.id).subscribe(
+        next => {
+          this.dataChangedEvent.emit();
+          this.router.navigate(['admin', 'users']);
+        }, error => this.message = 'Sorry, this user cannot be deleted at this time.'
+      );
+    }
   }
 
   resetPassword() {
-    this.dataService.resetUserPassword(this.user.id).subscribe();
+    this.message = 'please wait...';
+    this.dataService.resetUserPassword(this.user.id).subscribe(
+      next => this.message = 'The password has been reset.',
+      error => this.message = 'Sorry, something went wrong.'
+    );
   }
 
 }
